Put the tx option object in its own file
