+++ blockchain/dkg/dkgtasks/dkg_task.go
@@ -3,8 +3,6 @@ package dkgtasks
 import (
 	"github.com/MadBase/MadNet/blockchain/objects"
 	"github.com/ethereum/go-ethereum/common"
-	"math/big"
-	"strings"
 )
 
 type ExecutionData struct {
@@ -13,30 +11,13 @@ type ExecutionData struct {
 	State          *objects.DkgState
 	Success        bool
 	StartBlockHash common.Hash
-	TxOpts         *TxOpts
-}
-
-type TxOpts struct {
-	TxHashes     []common.Hash
-	Nonce        *big.Int
-	GasFeeCap    *big.Int
-	GasTipCap    *big.Int
-	MinedInBlock uint64
+	TxOpts         *objects.TxOpts
 }
 
 func (d *ExecutionData) Clear() {
-	d.TxOpts = &TxOpts{
+	d.TxOpts = &objects.TxOpts{
 		TxHashes: make([]common.Hash, 0),
 	}
-}
-
-func (t *TxOpts) GetHexTxsHashes() string {
-	var hashes strings.Builder
-	for _, txHash := range t.TxHashes {
-		hashes.WriteString(txHash.Hex())
-		hashes.WriteString(" ")
-	}
-	return hashes.String()
 }
 
 func NewExecutionData(state *objects.DkgState, start uint64, end uint64) *ExecutionData {
@@ -45,6 +26,6 @@ func NewExecutionData(state *objects.DkgState, start uint64, end uint64) *Execut
 		Start:   start,
 		End:     end,
 		Success: false,
-		TxOpts:  &TxOpts{TxHashes: make([]common.Hash, 0)},
+		TxOpts:  &objects.TxOpts{TxHashes: make([]common.Hash, 0)},
 	}
 }
